refactor(variable-value-editor-modal): simplify value editor logic and remove unused functions
